feat(bbapi): add flavor dispatch for stateful proving key caching
- Create bbapi_flavor_dispatch.hpp with UltraFlavorType enum
- Add select_ultra_flavor() function based on ProofSystemSettings
- Extract get_proving_key<Flavor> template in bbapi_stateful.hpp
- Refactor AcirGetProvingKey to dispatch to all Ultra* flavors
- Refactor AcirProveWithPk to dispatch to all Ultra* flavors
- Support Ultra, UltraZK, UltraKeccak, UltraKeccakZK, UltraRollup
