+++ barretenberg/cpp/src/barretenberg/bbapi/bbapi_ultra_honk.cpp
@@ -1,5 +1,7 @@
 #include "barretenberg/bbapi/bbapi_ultra_honk.hpp"
+#include "barretenberg/bbapi/bbapi_flavor_dispatch.hpp"
 #include "barretenberg/bbapi/bbapi_shared.hpp"
+#include "barretenberg/bbapi/bbapi_stateful.hpp"
 #include "barretenberg/circuit_checker/circuit_checker.hpp"
 #include "barretenberg/commitment_schemes/ipa/ipa.hpp"
 #include "barretenberg/common/log.hpp"
@@ -415,108 +417,36 @@ CircuitWriteSolidityVerifier::Response CircuitWriteSolidityVerifier::execute(BB_
     return { std::move(contract) };
 }
 
-/**
- * @brief Serializable polynomial data for proving key export.
- * @details Captures all metadata needed to reconstruct a polynomial:
- *   - coefficients: The actual coefficient data
- *   - start_index: The starting index of the memory-backed range (used for shifts)
- *   - virtual_size: The total logical size of the polynomial
- *
- * This allows proper reconstruction of "shiftable" polynomials that require
- * start_index > 0 for the shifted() operation to work correctly.
- */
-struct PolynomialExport {
-    std::vector<bb::fr> coefficients;
-    uint64_t start_index;
-    uint64_t virtual_size;
-
-    MSGPACK_FIELDS(coefficients, start_index, virtual_size);
-};
-
-/**
- * @brief Serializable proving key data for UltraFlavor.
- * @details Contains circuit-specific precomputed data that can be
- * cached and reused across multiple proofs with different witnesses.
- *
- * Key design decisions for upstream compatibility:
- * 1. PolynomialExport preserves start_index/virtual_size for proper reconstruction
- * 2. Bytecode hash enables cache validation across circuit versions
- * 3. All integer types use uint64_t for cross-platform msgpack compatibility
- */
-struct DeciderProvingKeyExport {
-    std::vector<PolynomialExport> polynomials;
-    std::vector<bb::fr> public_inputs;
-    bb::RelationParameters<bb::fr> relation_parameters;
-    std::vector<bb::fr> gate_challenges;
-    bb::fr target_sum;
-    bool is_structured;
-    uint64_t dyadic_size;
-    uint64_t num_public_inputs;
-    uint64_t pub_inputs_offset;
-    uint64_t overflow_size;
-    uint64_t final_active_wire_idx;
-    // Bytecode hash for cache validation - ensures proving key matches circuit
-    std::vector<uint8_t> bytecode_hash;
-    // Memory records for RAM/ROM lookup arguments (indices into full trace)
-    std::vector<uint32_t> memory_read_records;
-    std::vector<uint32_t> memory_write_records;
-
-    MSGPACK_FIELDS(polynomials,
-                   public_inputs,
-                   relation_parameters,
-                   gate_challenges,
-                   target_sum,
-                   is_structured,
-                   dyadic_size,
-                   num_public_inputs,
-                   pub_inputs_offset,
-                   overflow_size,
-                   final_active_wire_idx,
-                   bytecode_hash,
-                   memory_read_records,
-                   memory_write_records);
-};
-
 AcirGetProvingKey::Response AcirGetProvingKey::execute(BB_UNUSED const BBApiRequest& request) &&
 {
     BB_BENCH_NAME(MSGPACK_SCHEMA_NAME);
-    using ProverInstance = ProverInstance_<UltraFlavor>;
 
-    // Compute bytecode hash BEFORE consuming bytecode (for cache validation)
-    auto bytecode_hash_arr = blake3::blake3s(circuit.bytecode);
-    std::vector<uint8_t> bytecode_hash_vec(bytecode_hash_arr.begin(), bytecode_hash_arr.end());
+    // Select appropriate flavor based on settings
+    auto flavor_type = select_ultra_flavor(settings);
 
-    // Build proving key from circuit
-    auto prover_instance = [&] {
-        const acir_format::ProgramMetadata metadata{};
-        acir_format::AcirProgram program{ acir_format::circuit_buf_to_acir_format(std::move(circuit.bytecode)) };
-        auto builder = acir_format::create_circuit<UltraCircuitBuilder>(program);
-        return std::make_shared<ProverInstance>(builder);
-    }();
-
-    // Extract ONLY precomputed polynomials (selectors, sigmas, ids, tables, lagrange)
-    // WitnessEntities (w_l, w_r, w_o, w_4, z_perm, lookup_inverses, etc.) are witness-dependent
-    // and must be computed fresh for each proof with the new witness
     DeciderProvingKeyExport export_data;
-    for (auto& poly : prover_instance->polynomials.get_precomputed()) {
-        PolynomialExport poly_export;
-        poly_export.coefficients = std::vector<bb::fr>(poly.data(), poly.data() + poly.size());
-        poly_export.start_index = poly.start_index();
-        poly_export.virtual_size = poly.virtual_size();
-        export_data.polynomials.push_back(std::move(poly_export));
-    }
-    export_data.public_inputs = prover_instance->public_inputs;
-    export_data.relation_parameters = prover_instance->relation_parameters;
-    export_data.gate_challenges = prover_instance->gate_challenges;
-    export_data.dyadic_size = prover_instance->dyadic_size();
-    export_data.num_public_inputs = prover_instance->num_public_inputs();
-    export_data.pub_inputs_offset = prover_instance->pub_inputs_offset();
-    export_data.final_active_wire_idx = prover_instance->get_final_active_wire_idx();
-    export_data.bytecode_hash = std::move(bytecode_hash_vec);
-    export_data.memory_read_records = prover_instance->memory_read_records;
-    export_data.memory_write_records = prover_instance->memory_write_records;
 
-    // Serialize to msgpack
+    // Dispatch to the correct flavor
+    switch (flavor_type) {
+    case UltraFlavorType::Ultra:
+        export_data = get_proving_key<UltraFlavor>(circuit, settings);
+        break;
+    case UltraFlavorType::UltraZK:
+        export_data = get_proving_key<UltraZKFlavor>(circuit, settings);
+        break;
+    case UltraFlavorType::UltraKeccak:
+        export_data = get_proving_key<UltraKeccakFlavor>(circuit, settings);
+        break;
+    case UltraFlavorType::UltraKeccakZK:
+        export_data = get_proving_key<UltraKeccakZKFlavor>(circuit, settings);
+        break;
+    case UltraFlavorType::UltraRollup:
+        // UltraRollup uses same proving key structure as Ultra
+        export_data = get_proving_key<UltraRollupFlavor>(circuit, settings);
+        break;
+    }
+
+    // Serialize proving key to msgpack
     msgpack::sbuffer buffer;
     msgpack::pack(buffer, export_data);
 
@@ -525,104 +455,31 @@ AcirGetProvingKey::Response AcirGetProvingKey::execute(BB_UNUSED const BBApiRequ
 
 AcirProveWithPk::Response AcirProveWithPk::execute(BB_UNUSED const BBApiRequest& request) &&
 {
+    BB_BENCH_NAME(MSGPACK_SCHEMA_NAME);
+
+    // Select appropriate flavor based on settings
+    auto flavor_type = select_ultra_flavor(settings);
+
     std::vector<uint8_t> result_vec;
-    {
-        BB_BENCH_NAME(MSGPACK_SCHEMA_NAME);
-        using ProverInstance = ProverInstance_<UltraFlavor>;
-        using VerificationKey = UltraFlavor::VerificationKey;
 
-        // Compute bytecode hash for cache validation
-        auto current_bytecode_hash = blake3::blake3s(circuit.bytecode);
-
-        // Deserialize proving key
-        DeciderProvingKeyExport pk_data;
-        msgpack::object_handle oh = msgpack::unpack((const char*)proving_key.data(), proving_key.size());
-        msgpack::object obj = oh.get();
-        obj.convert(pk_data);
-
-        // Validate bytecode hash matches proving key
-        bool hash_matches =
-            (pk_data.bytecode_hash.size() == current_bytecode_hash.size()) &&
-            std::equal(pk_data.bytecode_hash.begin(), pk_data.bytecode_hash.end(), current_bytecode_hash.begin());
-
-        if (!hash_matches) {
-            throw_or_abort("AcirProveWithPk: Bytecode hash mismatch. "
-                           "The proving key was generated for a different circuit. "
-                           "Please regenerate the proving key with the current bytecode.");
-        }
-
-        // Reconstruct circuit from bytecode and witness
-        acir_format::AcirProgram program{ acir_format::circuit_buf_to_acir_format(std::move(circuit.bytecode)) };
-        program.witness = acir_format::witness_buf_to_witness_vector(std::move(witness));
-        auto builder = acir_format::create_circuit<UltraCircuitBuilder>(program);
-
-        // Reconstruct metadata from proving key
-        MetaData metadata;
-        metadata.dyadic_size = static_cast<size_t>(pk_data.dyadic_size);
-        metadata.num_public_inputs = static_cast<size_t>(pk_data.num_public_inputs);
-        metadata.pub_inputs_offset = static_cast<size_t>(pk_data.pub_inputs_offset);
-
-        // Convert PolynomialExport to ProverInstance::PolynomialData
-        std::vector<ProverInstance::PolynomialData> poly_data_vec;
-        poly_data_vec.reserve(pk_data.polynomials.size());
-        for (auto& poly_export : pk_data.polynomials) {
-            ProverInstance::PolynomialData poly_data;
-            poly_data.coefficients = std::move(poly_export.coefficients);
-            poly_data.start_index = static_cast<size_t>(poly_export.start_index);
-            poly_data.virtual_size = static_cast<size_t>(poly_export.virtual_size);
-            poly_data_vec.push_back(std::move(poly_data));
-        }
-
-        // HYDRATION: Create ProverInstance using precomputed polynomials from proving key
-        // This skips recomputing selectors, permutation polynomials, and lookup tables
-        auto instance = std::make_shared<ProverInstance>(builder,
-                                                         std::move(poly_data_vec),
-                                                         metadata,
-                                                         static_cast<size_t>(pk_data.final_active_wire_idx),
-                                                         std::move(pk_data.memory_read_records),
-                                                         std::move(pk_data.memory_write_records));
-
-        // Construct verification key and prove
-        auto verification_key = std::make_shared<VerificationKey>(instance->get_precomputed());
-        UltraProver prover{ instance, verification_key };
-
-        auto proof = prover.construct_proof();
-
-        // Calculate inner public inputs (excluding pairing point accumulator), consistent with CircuitProve
-        size_t num_inner_public_inputs = [&]() {
-            size_t num_public_inputs = instance->num_public_inputs();
-            // UltraFlavor uses DefaultIO::PUBLIC_INPUTS_SIZE for pairing point accumulator
-            constexpr size_t PAIRING_POINT_SIZE = 16; // DefaultIO::PUBLIC_INPUTS_SIZE
-            BB_ASSERT(num_public_inputs >= PAIRING_POINT_SIZE,
-                      "Public inputs should contain a pairing point accumulator.");
-            return num_public_inputs - PAIRING_POINT_SIZE;
-        }();
-
-        // Create the combined result
-        // Format: [num_public_inputs (4 bytes)] [public_inputs (32 bytes each)] [proof...]
-        size_t total_size = 4 + (num_inner_public_inputs * 32) + (proof.size() * 32);
-        result_vec.resize(total_size);
-
-        uint8_t* ptr = result_vec.data();
-
-        // Pack num_inner_public_inputs (4 bytes, big endian, matches CircuitProve format)
-        uint32_t num_pub_inputs_be = htonl(static_cast<uint32_t>(num_inner_public_inputs));
-        std::memcpy(ptr, &num_pub_inputs_be, 4);
-        ptr += 4;
-
-        // Pack inner public inputs (from proof, excluding pairing point accumulator)
-        // These are the first num_inner_public_inputs elements of the proof
-        for (size_t i = 0; i < num_inner_public_inputs; ++i) {
-            bb::fr::serialize_to_buffer(proof[i], ptr);
-            ptr += 32;
-        }
-
-        // Pack proof (skip public inputs which were already extracted above)
-        for (size_t i = num_inner_public_inputs; i < proof.size(); ++i) {
-            bb::fr::serialize_to_buffer(proof[i], ptr);
-            ptr += 32;
-        }
-    } // Destructors run here
+    // Dispatch to the correct flavor
+    switch (flavor_type) {
+    case UltraFlavorType::Ultra:
+        result_vec = prove_with_pk<UltraFlavor>(circuit, witness, proving_key, settings);
+        break;
+    case UltraFlavorType::UltraZK:
+        result_vec = prove_with_pk<UltraZKFlavor>(circuit, witness, proving_key, settings);
+        break;
+    case UltraFlavorType::UltraKeccak:
+        result_vec = prove_with_pk<UltraKeccakFlavor>(circuit, witness, proving_key, settings);
+        break;
+    case UltraFlavorType::UltraKeccakZK:
+        result_vec = prove_with_pk<UltraKeccakZKFlavor>(circuit, witness, proving_key, settings);
+        break;
+    case UltraFlavorType::UltraRollup:
+        result_vec = prove_with_pk<UltraRollupFlavor>(circuit, witness, proving_key, settings);
+        break;
+    }
 
     return { .combined_result = std::move(result_vec) };
 }
